requirements.txt
# Python dependencies for MTA Transit Data Analytics Dashboard

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# API requests
requests>=2.31.0
urllib3>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Logging and utilities
colorlog>=6.7.0

# Data validation
pydantic>=2.0.0

# Date/time utilities
python-dateutil>=2.8.0

# Progress bars (optional)
tqdm>=4.65.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
